Make Plang case-insensitive
So, you can register stuff in Context by a Name object. The Name is
really just a marker interface: implementors should implement hashCode
and equals so as to meet their needs.

In PlangGrammar Name holds the case-insensitive version of the
token.lexeme found in the source. Other grammars may use the Name to
take only the first six characters of the name into account or
something.

It's fucking elegant, if I may say so myself