AST nodes keep location information
This was painful, but it would have been even more painful if left for
later.

So, Nodes know where they were in the source code. It's kinda shaky and
uncertain yet. Assignments, for example, their location is currently the
location of the := sign, while the location of conditional statements is
their whole block, everything between (and including) HA and HA_VÉGE.
This is inconsistent as fuck, but one could argue that if there's a
runtime problem with the assignment, we'd want the interpreter to point
at the := sign instead of the entire line.

... which kinda means assignment is an operator, after all.

Also, maybe plang.SequenceParser should be a front-testing loop.