Make PLanG have its own types
This was a bit heavy, and painful, and a bit useless, but here we go
anyway: the Grammar is supposed to create its own types.

The problem with using the Archetypes was really only the labels. When
reporting errors, it used IntegerType etc, while in fact the types had
names defined in the lexical.properties. I toyed around with a number of
solutions, but none seemed satisfactory.

Things would be a lot simpler if we hardcoded the lexical properties. Do
we really need an English PLanG?