Factor out reader logic to own class
That also contains the name of the stream. It's better this way for the
separation of concerns

InputStreamValue is now concerned only with the creation of new values.
I'm not terribly sure about this, but I like how the "if
the returned value is null, then return a nullValue" logic is
encapsulated there.