Make it possible to compile and run from the GUI
It's a bare-bones implementation: errors are not yet shown, the
interpreter runs in the console.

Compiler and Executor are simple, stateless classes. The compilation
result is kept by the MainFrame. Compiler may eventually be refactored
into a SwingWorker, though compilation is so quick it doesn't really
need to.

The Executor will need some mechanisms to exchange data with other
components, and it will definitely need to be asynchronous, otherwise
it'd lock up the GUI while waiting for input.