Start the editor as a separate program
Starting it from the command line with the --gui option was nice, but
all the other command line options were incompatible with it. The
cleanest solution is to have it as a separate binary, as it were.

The problem with the new startup script on Windows is, it momentarily opens the
command line window (as it's a batch script). We'll see about that