Refactor Parser error reporting
Gone is the beautiful exception hierarchy! We NEVER caught those
exceptions or differentiated them by type anyway.

Gone is the Translator! It was ugly, and forced Parser to be aware of
the current locale.

Now we have an ErrorCode based system. The number of error codes may end
up pretty large, though